+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2022-07-01
+### Added
+- Docker utils
+
+## [0.3.0] - 2022-07-01
+### Added
+- Storage API
+
+## [0.2.1] - 2022-07-01
+### Fixed
+- Sidecars bugs
+
+## [0.2.0] - 2022-06-30
+### Added
+- Sidecars for pings and heartbeats
+
 ## [0.1.0] - 2022-06-10
 ### Added
 - Initial commit with basic structure